xpath: Move expression AST into a seperate crate
